HUE edits

See merge request cse1105/2022-2023/teams/oopp-team-45!59

+++ client/src/main/java/client/scenes/MainSceneCtrl.java
@@ -7,9 +7,11 @@ import com.google.inject.Inject;
 import commons.Board;
 import commons.TaskList;
 import javafx.fxml.FXML;
-import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 
+import javafx.scene.control.MenuItem;
+import javafx.scene.image.Image;
+import javafx.scene.image.ImageView;
 import javafx.scene.input.Clipboard;
 import javafx.scene.input.ClipboardContent;
 import javafx.scene.layout.HBox;
@@ -34,12 +36,29 @@ public class MainSceneCtrl {
     VBox boardsContainer;
     @FXML
     HBox taskListsContainer;
+
     @FXML
-    Button renameBoard;
+    MenuItem createBoardMenu;
     @FXML
-    Button removeBoard;
+    MenuItem renameBoardMenu;
     @FXML
-    Button copyCode;
+    MenuItem deleteBoardMenu;
+    @FXML
+    MenuItem joinBoardMenu;
+    @FXML
+    MenuItem joinServerMenu;
+
+    @FXML
+    ImageView menuIcon;
+    @FXML
+    ImageView adminIcon;
+    @FXML
+    ImageView copyIcon;
+
+    @FXML
+    Label boardCode;
+    @FXML
+    Label serverAddr;
 
     /**
      * constructor
@@ -67,6 +86,18 @@ public class MainSceneCtrl {
                 "TaskList.fxml"
         );
 
+        Image menu = new Image(getClass()
+                .getResourceAsStream("/client/images/menuicon.png"));
+        menuIcon.setImage(menu);
+
+        Image admin = new Image(getClass()
+                .getResourceAsStream("/client/images/adminicon.png"));
+        adminIcon.setImage(admin);
+
+        Image copy = new Image(getClass()
+                .getResourceAsStream("/client/images/copyicon.png"));
+        copyIcon.setImage(copy);
+
         // Set default board as current board (needs FXML title)
         setActiveBoard(server.getDefaultBoard());
     }
@@ -85,6 +116,7 @@ public class MainSceneCtrl {
     public void setActiveBoard(Board activeBoard) {
         this.activeBoard = activeBoard;
         sceneTitle.setText(activeBoard.getTitle());
+        boardCode.setText(activeBoard.getCode());
 
         refresh();
     }
@@ -139,8 +171,14 @@ public class MainSceneCtrl {
             ErrorUtils.alertError("You cannot delete the default board!");
             return;
         }
-        server.deleteBoard(activeBoard);
-        setActiveBoard(server.getDefaultBoard());
+
+        boolean confirmation = server.confirmDeletion("board");
+
+        // Check the user's response and perform the desired action
+        if (confirmation) {
+            server.deleteBoard(activeBoard);
+            setActiveBoard(server.getDefaultBoard());
+        }
     }
 
     /**
@@ -158,6 +196,14 @@ public class MainSceneCtrl {
 
         System.out.println("The code for this board is copied!");
         System.out.println("Code: " + code);
+    }
+
+    /**
+     * set the server address to be displayed
+     * @param address the address to be displayed
+     */
+    public void setServerAddr(String address) {
+        this.serverAddr.setText(address);
     }
 
     /**
